requirements.txt
fastapi
uvicorn
langchain
pinecone-client
pdfplumber
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
requests
python-dotenv 